pyproject.toml
[project]
name = "alchecode"
version = "0.1.0"
description = ""
authors = [
    {name = "tbphp",email = "admin@tbphp.net"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
]

[tool.mypy]
strict = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D415",
    "D417",
    "E501",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.poetry]
name = "alchecode"
packages = [
    {include = "alchecode", from = "src"},
    {include = "config", from = "src"},
    {include = "utils", from = "src"},
]

[tool.poetry.scripts]
"acode:cli" = "alchecode:cli"
"acode:api" = "alchecode:api_server"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"
mypy = "^1.15.0"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
